% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amce.r
\name{autoplot.amce}
\alias{autoplot.amce}
\title{Plot Average Marginal Component Effects (AMCEs) from a Conjoint Experiment}
\usage{
autoplot.amce(df)
}
\arguments{
\item{df}{An object of class \code{amce}, as returned by \code{\link{amce}}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Produces a dot-and-whisker plot of AMCE estimates, with one row per
attribute level. Error bars represent +/- one standard error. Reference
levels are plotted at zero with no error bar. Points are colored by
attribute.
}
\examples{
amces <- amce(data, selected ~ group + sex + age, id = ~uuid)
autoplot(amces)

}
\seealso{
\code{\link{amce}}, \code{\link{autoplot.marginal_means}}
}
