% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_means.r
\name{autoplot.marginal_means}
\alias{autoplot.marginal_means}
\title{Plot Marginal Means from a Conjoint Experiment}
\usage{
autoplot.marginal_means(df)
}
\arguments{
\item{df}{An object of class \code{marginal_means}, as returned by
\code{\link{marginal_means}}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Produces a dot-and-whisker plot of marginal means estimates, with one row
per attribute level. Error bars represent +/- one standard error. Points
are colored by attribute.
}
\examples{
mms <- marginal_means(data, selected ~ group + sex + age, cluster = ~uuid)
autoplot(mms)

}
\seealso{
\code{\link{marginal_means}}
}
