% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_means.r
\name{conditional_marginal_means}
\alias{conditional_marginal_means}
\title{Estimate Conditional Marginal Means for a Conjoint Experiment}
\usage{
conditional_marginal_means(
  data,
  formula = NULL,
  outcome = NULL,
  attributes = NULL,
  id = NULL,
  group = NULL,
  vcov_type = "HC1",
  wts = NULL,
  design = NULL
)
}
\arguments{
\item{data}{A data frame containing the conjoint data}

\item{formula}{A formula of the form \code{outcome ~ attr1 + attr2 + ...}}

\item{outcome}{(Alternative to \code{formula}) Character string naming the outcome variable}

\item{attributes}{(Alternative to \code{formula}) Character vector of attribute names}

\item{id}{(Optional) A one-sided formula specifying the ID/cluster variable, e.g. \code{~id}}

\item{group}{The respondent-level grouping variable (unquoted). Marginal
means are estimated separately for each level of this variable.}

\item{vcov_type}{Type of variance-covariance estimation when clustering (HC0-HC3). Default is "HC1".}

\item{wts}{(Optional) Weights to be used in the regression. Can be
\code{NULL} (the default), a numeric vector, or the name of a
column in \code{data} (quoted or unquoted).}

\item{design}{A \code{survey::svydesign}-object. If a \code{design} is
provided, \code{cjlm} uses \code{survey::svyglm} as backend using
the provided design, disregarding other arguments like \code{id},
\code{vcov_type}, and \code{wts}, as all of these are handled
by \code{survey::svyglm} (see \code{?survey::svyglm}).}
}
\value{
A data frame of class \code{marginal_means}
}
\description{
Conditional version of \code{\link{marginal_means}}. Computes marginal means separately for each
level of a respondent-level grouping variable, allowing comparisons of conjoint preferences across
subgroups. This approach is preferred over conditional AMCEs for subgroup analysis, as conditional
marginal means do not require a baseline, and are thus not sensitive to the choice of a baseline
category. This was recommended by Leeper, Hobolt & Tilley (2020).
}
\details{
By default, the function uses \code{stats::lm} to estimate the model, and
\code{sandwich} to adjust standard errors if needed. However, you can also pass
a surveydesign (\code{survey::svydesign}), in which case \code{survey::svyglm}
will be used.
}
\examples{
marginal_means(data, selected ~ group + sex + age, id = ~id)
}
\references{
Leeper, T. J., Hobolt, S. B., and Tilley, J. (2020). Measuring
Subgroup Preferences in Conjoint Experiments. \emph{Political Analysis},
28(2), 207--221. \doi{10.1017/pan.2019.30}
}
