% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_means.r
\name{marginal_means}
\alias{marginal_means}
\title{Estimate Marginal Means for a Conjoint Experiment}
\usage{
marginal_means(data, formula = NULL, outcome = NULL, attributes = NULL, id)
}
\arguments{
\item{data}{A data frame containing the conjoint data}

\item{formula}{A formula of the form \code{outcome ~ attr1 + attr2 + ...}}

\item{outcome}{(Alternative to \code{formula}) Character string naming the outcome variable}

\item{attributes}{(Alternative to \code{formula}) Character vector of attribute names}

\item{cluster}{A one-sided formula specifying the ID/cluster variable, e.g. \code{~id}}
}
\value{
A data frame of class \code{marginal_means}
}
\description{
Estimate Marginal Means for a Conjoint Experiment
}
\examples{
marginal_means(data, selected ~ group + sex + age, id = ~id)
}
